更改逻辑删除查询bug

+++ src/main/java/com/xiliulou/electricity/controller/admin/JsonAdminElectricityPayParamsController.java
@@ -71,7 +71,7 @@ public class JsonAdminElectricityPayParamsController extends AbstractFranchiseeD
     
     @DeleteMapping(value = "/admin/electricityPayParams/{id}")
     @Log(title = "删除支付参数")
-    public R delete(@PathVariable("id") Long id) {
+    public R delete(@PathVariable("id") Integer id) {
         return electricityPayParamsService.delete(id);
     }
